+++ bitwise/processor/__init__.py
@@ -1,3 +1,4 @@
 from .ALU import *
-from .BUS import *
+from .FLAG import *
 from .PROC import *
+from .SP import *