+++ tests/processor/test_Processor.py
@@ -1,4 +1,0 @@
-import bitwise as bw
-
-
-# class TestProcessor: